LPS-112426 Loads the resources of the field types to the ReactFieldAdapter
