+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/resources/META-INF/resources/js/components/AutoFocus.es.js
@@ -25,7 +25,7 @@ export const AutoFocus = ({children}) => {
 	const lazyStatus = children?.props?.children?.type?._status;
 
 	useLayoutEffect(() => {
-		if (childRef.current && containerElement.current) {
+		if (childRef.current && containerElement && containerElement.current) {
 			const firstInput = childRef.current.querySelector('input');
 
 			if (
